requirements.txt
python-dotenv
langchain-groq
langchain-core
langchain-community
chromadb
pandas
streamlit
